refactoring in progress 4
