+++ android/src/main/java/io/parity/signer/screens/onboarding/WaitingScreen.kt
